Dedupe quaternion helpers, wire up dead z offset, fix minor naming
- Move _quaternion_multiply/_conjugate/_axis_angle/_wrap_angle/
  _quaternion_from_euler out of centerpose_box.py and centerpose_shoe.py
  into shared PickPlaceNodeBase (verbatim, was byte-identical in both).
- Apply grasp_position_offset[2] (z) to the computed grasp pose in
  bottle/box/shoe instead of silently dropping it; behavior-neutral
  today since every default/launch config has z=0.0.
- Note in centerpose_box.py that grasp_position_x_slope intentionally
  reuses grasp_position_y_reference_pixel as the shared reference pixel.
- Give centerpose_shoe.py a named _LOCAL_Y_180_FLIP constant matching
  centerpose_box.py instead of an inline literal.
